refactor: 불필요한 view 패키지 및 view 클래스 삭제 및 dto 패키지 이동
