+++ src/main/java/chess/dto/ScoreDto.java
@@ -1,4 +1,4 @@
-package chess.view.dto;
+package chess.dto;
 
 import chess.domain.game.ChessGame;
