--- src/main/java/chess/view/InputView.java
@@ -1,16 +1,0 @@
-package chess.view;
-
-import java.util.Scanner;
-
-public class InputView {
-
-    private static final Scanner SCANNER = new Scanner(System.in);
-
-    private InputView() {
-    }
-
-    public static String inputCommandFromUser() {
-        return SCANNER.nextLine();
-    }
-
-}
